refactor(drift_sync): extract checks.py (5/8)
Move check_labels, check_protection, check_profile_files and helpers
_read_template_content, _content_hash into checks.py. Each @register_check
fires when checks.py is imported by __init__.py.

Module-attribute bindings (labels_api, protection_api, issues_api) are
preserved in __init__.py so test mocker.patch paths continue to work.
Because Python module objects are singletons, the binding inside
checks.py and the binding in __init__.py refer to the SAME object —
patching either path flows through.

Verified: drift_sync.labels_api is gh_manage.github_api.labels (identity);
_CHECKS populated with 4 checks (3 drift + check_shape from doctor bridge).

Unused imports (hashlib, Path, _package_files) cleaned from __init__.py.

Refs #47 (Theme A item 4).
